Initial form processor refactor
