+++ .idea/fmnas.iml
@@ -3,10 +3,7 @@
   <component name="Go" enabled="true" />
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/admin" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/lib" isTestSource="false" packagePrefix="PHPMailer" />
-      <sourceFolder url="file://$MODULE_DIR$/public" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/admin_old" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/Form" isTestSource="false" packagePrefix="fmnas\Form\" />
       <excludeFolder url="file://$MODULE_DIR$/vendor/erusev/parsedown" />
       <excludeFolder url="file://$MODULE_DIR$/vendor/phpmailer/phpmailer" />
       <excludeFolder url="file://$MODULE_DIR$/vendor/composer" />
