+++ src/.gitignore
@@ -1,4 +1,1 @@
 generated.php
-*/
-!errors/
-!templates/
